+++ terraform/modules/ingest_pipeline/outputs.tf
@@ -21,3 +21,11 @@ output "ingest_lambda_role_arn" {
 output "ingest_lambda_role_name" {
   value = aws_iam_role.ingest_lambda.name
 }
+
+output "ingest_lambda_arn" {
+  value = aws_lambda_function.ingest.arn
+}
+
+output "ingest_lambda_function_name" {
+  value = aws_lambda_function.ingest.function_name
+}